requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv==1.0.0
PyPDF2==3.0.1

# Updated LangChain packages (modern approach)
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20

# Vector store and embeddings
faiss-cpu>=1.7.4

# Additional dependencies
tiktoken>=0.5.0
openai>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Web framework dependencies (will be installed with streamlit)
altair>=4.0.0
